Fix error for returning user amending address

+++ vulnerable_people_form/form_pages/postcode_lookup.py
@@ -5,7 +5,12 @@ from .blueprint import form
 from .shared.querystring_utils import append_querystring_params
 from .shared.render import render_template_with_title
 from .shared.routing import route_to_next_form_page
-from .shared.session import get_errors_from_session, request_form, delete_answer_from_form
+from .shared.session import (
+    get_errors_from_session,
+    request_form,
+    delete_answer_from_form,
+    accessing_saved_answers
+)
 from .shared.validation import validate_postcode
 
 
@@ -15,7 +20,9 @@ def post_postcode_lookup():
     if not validate_postcode(session["postcode"], "postcode"):
         return redirect("/postcode-lookup")
 
-    delete_answer_from_form("support_address")
+    if not accessing_saved_answers():
+        delete_answer_from_form("support_address")
+
     session["error_items"] = {}
     return route_to_next_form_page()
